# Simulation performance improvements