@kottmanj 2024-08-27
some of the code you changed is still from the first prototype of tequila -- back then there wasn't much thought on speed. Would have make sense to improve it earlier, but I think it's because the effects on the variational applications are not that high. The wfn = tq.simulate(U) wavefunction simulation was initially just a bonus feature (as tq it is primarily designed for expectation values) -- But it is great to have that bonus feature accelerated :-) [especially since we also often use it for development purposes]

I need to double check code changes, but I think this is good to go.

Thank you once more!